phy_device: force use of recovered clock
Force generic phy driver to use 100 Base-T recovered clock
instead of the local clock.

This is a bandaid fix that works around a HW clock error.
The local clock is out of spec enough to cause errors, while
the recovered clock is still correct and can be used instead.

Since we are using the generic phy driver, we can't parameterize
this particular register setup from the dts.